Completed and clarifying exercise 1
This contains a method that uses string interpolation to read and
ultimately print a string.  I use the double quotes which creates a
string literal so that the string interpolation and the escape
characters will work as expected.  This prints one line at a time
because of the \n character being recognized, so it does have a line
return at these places; however, it does not pause in between printing
out each of the three lines.  Should it?