Figured out exercise 1 by investigating all the string methods
I went to www.ruby-doc.org and read through all of the names of the
methods, and did trial and error over and over with a bunch of methods
until I found something that showed a glimmer of hope.  This was the
split method.  I then remembered the prepend method that I had come
across.  I combined those two with the an iterator loop (which we have
not learned about yet, I know), and was able to get the desired
results.